Installer can now create links on desktop and start menu (Windows)

+++ src/Winecrash/Installer/Installer.cs
@@ -1,6 +1,8 @@
 ﻿using System;
+using System.ComponentModel;
 using System.IO;
 using System.IO.Compression;
+using System.Net;
 using System.Windows.Forms;
 using System.Reflection;
 
@@ -8,6 +10,8 @@ namespace Winecrash.Installer
 {
     public partial class Installer : Form
     {
+        private WebClient _client = new WebClient();
+        
         public Installer()
         {
             InitializeComponent();
@@ -15,33 +19,76 @@ namespace Winecrash.Installer
 
         private void Installer_Load(object sender, EventArgs e)
         {
-            InstallLauncher(Utilities.DefaultInstallDirectory);
+            _pathInput.Text = Utilities.DefaultInstallDirectory;
+            _client.DownloadFileCompleted += _client_InstallFileCompleted;
+            _client.DownloadProgressChanged += _client_DownloadProgress;
         }
 
         public void InstallLauncher(string path)
         {
-            /*string[] names = Assembly.GetExecutingAssembly().GetManifestResourceNames();
-
-            foreach (string name in names)
-            {
-                MessageBox.Show(name);
-            }*/
-            
             Utilities.SetupForInstallation(path);
-            string zipPath = Path.Combine(Utilities.ApplicationPath, "temp.zip");
+            
+            _cancelButton.Enabled = true;
+            
+            _client.DownloadFileAsync(new Uri(Utilities.DownloadLink), Utilities.ZipPath);
+        }
 
-            using (Stream cs = Assembly.GetExecutingAssembly().GetManifestResourceStream(Utilities.EmbeddedFileName))
-            using (FileStream fs = new FileStream(zipPath, FileMode.Create)) cs.CopyTo(fs);
+        private void _pathInput_TextChanged(object sender, EventArgs e)
+        {
+            _actualInstallationLabel.Text = Path.Combine(_pathInput.Text, Utilities.ApplicationName);
+        }
+
+        private void _cancelButton_Click(object sender, EventArgs e)
+        {
+            _client.CancelAsync();
+        }
+
+        private void _client_DownloadProgress(object sender, DownloadProgressChangedEventArgs e)
+        {
+            _installProgress.Value = e.ProgressPercentage;
+        }
+
+        private void _client_InstallFileCompleted(object sender, AsyncCompletedEventArgs e)
+        {
+            _cancelButton.Enabled = false;
             
-            ZipFile.ExtractToDirectory(zipPath, Utilities.LauncherPath);
-            
-            File.Delete(zipPath);
-            /*using (var zipStream = new GZipStream(compressedStream, CompressionMode.Decompress))
-            using (var resultStream = new MemoryStream())
+            if (e.Cancelled)
             {
-                zipStream.CopyTo(resultStream);
-                zipStream.
-            }*/
+                MessageBox.Show("Installation Cancelled");
+            }
+            else
+            {
+                string zipPath = Utilities.ZipPath;
+
+                ZipFile.ExtractToDirectory(zipPath, Utilities.LauncherPath);
+
+                File.Delete(zipPath);
+                if (_createDesktopShortcut.Checked)
+                    Utilities.CreateShortcutDesktop(Utilities.LauncherExecutablePath);//Path.Combine(Utilities.LauncherPath, "Winecrash.exe"));
+                if (_createStartShortcut.Checked)
+                    Utilities.CreateShortcutStart(Utilities.LauncherExecutablePath);//Path.Combine(Utilities.LauncherPath, "Winecrash.exe"));
+            
+                
+            }
+        }
+
+        private void _buttonInstall_Click(object sender, EventArgs e)
+        {
+            string path = _pathInput.Text;
+            if (Utilities.InstalledAtPath(Path.Combine(path, Utilities.ApplicationName)))
+            {
+                DialogResult result =
+                    MessageBox.Show(
+                        $"{Utilities.ApplicationName} is already installed at this directory. Do you want to replace it?", "Directory Conflict", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+
+                if (result == DialogResult.Yes)
+                {
+                    Directory.Delete(Path.Combine(path, Utilities.ApplicationName), true);
+                    InstallLauncher(path);
+                }
+            }
+            else
+                InstallLauncher(path);
         }
     }
 }
